feat(programmes): allow half-point ECTS credits — supports 7.5-credit subjects
Subject.Ects was an int, so common values like 7.5 ECTS could not be
entered. It is now decimal, mapped numeric(5,1), with the ECTS input
accepting step 0.5. The subject create/update DTOs (admin CRUD, partner
programme update, public programmes) take decimal, and the transcript
renderers carry ECTS as decimal — grade-point maths casts to double and
ECTS prints via "0.##" so 7.5 shows as 7.5 and 6 as 6 (no trailing .0).

Migration AddSubjectEctsDecimal alters the column int→numeric(5,1);
existing whole values convert losslessly.

Affects: Subject, SubjectConfiguration, SubjectApi, MyPrograms update, PublicProgrammes, LetterTagResolver, LetterPdfRenderer, AdminAcademicView, Migrations
Breaking: None

+++ backend/Api/PublicSignupApi/SharedLibrary.Basics.Opaque.PublicSignupApi/V1/Programmes/Endpoint/PublicProgrammesV1ListEndpoint.cs
@@ -146,5 +146,5 @@ public sealed class PublicSubjectItem
     public required Guid SubjectId { get; init; }
     public required string Code { get; init; }
     public required string Name { get; init; }
-    public required int Ects { get; init; }
+    public required decimal Ects { get; init; }
 }
